+++ resources/views/site/view.php
@@ -11,8 +11,8 @@ declare(strict_types=1);
 
 use Hirtz\Cms\Controllers\SiteController;
 use Hirtz\Cms\Models\Entry;
-use Hirtz\Cms\widgets\MetaTags;
-use Hirtz\Cms\widgets\Sections;
+use Hirtz\Cms\Widgets\MetaTags;
+use Hirtz\Cms\Widgets\Sections;
 use Hirtz\Skeleton\Web\View;
 
 echo MetaTags::make()