+++ src/Controllers/SiteController.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hirtz\Cms\Controllers;
 
-use Hirtz\Cms\Models\builders\EntrySiteRelationsBuilder;
+use Hirtz\Cms\Models\Builders\EntrySiteRelationsBuilder;
 use Hirtz\Cms\Models\Entry;
 use Hirtz\Cms\Models\Queries\EntryQuery;
 use Hirtz\Cms\Module;
@@ -20,6 +20,7 @@ class SiteController extends Controller
 {
     public bool $redirectTrailingSlash = true;
 
+    #[\Override]
     public function init(): void
     {
         $this->layout ??= 'main';