+++ src/Modules/Admin/Widgets/Forms/Traits/MenuFieldTrait.php
@@ -4,13 +4,14 @@ declare(strict_types=1);
 
 namespace Hirtz\Cms\Modules\Admin\Widgets\Forms\Traits;
 
+use Hirtz\Cms\Models\Entry;
 use Hirtz\Cms\Models\Traits\MenuAttributeTrait;
 use Hirtz\Skeleton\Widgets\Forms\Fields\CheckboxField;
 use Stringable;
 use Yii;
 
 /**
- * @property MenuAttributeTrait $model
+ * @property Entry&MenuAttributeTrait $model
  */
 trait MenuFieldTrait
 {